Caching of enpoints [more in description]
For caching to work correctly, or rather the routing of FastAPI itself, it was necessary to add an optional value to the ArtworkReadSchema in the "created_at" field, which is excluded by default. If you do not do this, then when serializing data in FastAPI routing (\site-packages\fastapi.py\routing.py [291]), when retrieving cached data from Redis, an error occurs when the "created_at" field is missing.